📈 fuck yes

+++ src/swe/engine/HistoryState.java
@@ -5,9 +5,9 @@ import java.util.Calendar;
 final public class HistoryState {
     final public TickIdentifier tickIdentifier;
     final public MarketStatus status;
-    final public Float shareIndex;
+    final public double shareIndex;
 
-    public HistoryState(TickIdentifier tickIdentifier, MarketStatus status, Float shareIndex) {
+    public HistoryState(TickIdentifier tickIdentifier, MarketStatus status, double shareIndex) {
         this.tickIdentifier = tickIdentifier;
         this.status = status;
         this.shareIndex = shareIndex;
